requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.40
psycopg[binary]==3.2.7
python-jose[cryptography]==3.4.0
pydantic-settings==2.8.1
email-validator==2.2.0
pytest==8.3.5
pytest-cov==5.0.0
httpx==0.28.1